requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==3.0.1
multidict==6.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
psutil==6.0.0
PyMuPDF==1.24.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.14.0
